float to double

+++ lib_cuda/hitable_list.h
@@ -8,7 +8,7 @@ class hitable_list: public hitable
 public:
     __device__ hitable_list() {}
     __device__ hitable_list(hitable **l, int n) {list = l; list_size = n;}
-    __device__ virtual bool hit(const ray& r, double t_min, double t_max, hit_record& rec) const;
+    __device__ virtual bool hit(const ray& r, float t_min, float t_max, hit_record& rec) const;
 	__device__ ~hitable_list()
 	{
 		for(int i=0; i<list_size; ++i)
@@ -18,11 +18,11 @@ public:
     int list_size;
 };
 
-__device__ bool hitable_list::hit(const ray& r, double t_min, double t_max, hit_record& rec) const
+__device__ bool hitable_list::hit(const ray& r, float t_min, float t_max, hit_record& rec) const
 {
     hit_record temp_rec;
     bool hit_anything = false;
-    double closest_so_far = t_max;
+    float closest_so_far = t_max;
     for(int i = 0; i < list_size; i += 1)
     {
         if(list[i]->hit(r, t_min, closest_so_far, temp_rec))
